requirements.txt
PyQt5
openai>=1.0.0
google-generativeai

